Add support for compact_bbc format
The format cleans up the amount of bbcode tags needed so that the output
for most standard size albums will be less than the size of the bbc format.
The more tracks on the album the more size savings you will see.

This change also removes [url][/url] wrapping around the spectrograms in
the analyzer.py code so that each formatter can wrap that appropriately.

This change does not contain albumartistm, album and release date as part
of the compact_bbc summary yet, there are some metadata changes needed for
that.